chore(webhook-server): rename to event-collectors
