+++ event-collectors/.env.example
