+++ event-collectors/.envrc.example
