+++ event-collectors/Dockerfile
