+++ event-collectors/package.json
