+++ event-collectors/src/util.js
